Refactor signal handling logic
