+++ src/main/cpp/sim/main.cpp
@@ -24,50 +24,7 @@
 #include <exception>
 #include <iostream>
 #include <tclap/CmdLine.h>
-
-#ifndef HANDLE_SIGNALS
-// Handle signals unless otherwise specified. This is especially useful for debugging
-// segfaults that occur on a remote machine.
-#define HANDLE_SIGNALS 1
-#endif
-
-#if HANDLE_SIGNALS
-
-#include <signal.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <execinfo.h>
-#include <unistd.h>
-
-void handle_segfault(int sig)
-{
-	// This signal handler is based on the signal handler from this StackOverflow answer:
-	//
-	// http://stackoverflow.com/questions/77005/how-to-generate-a-stacktrace-when-my-gcc-c-app-crashes
-	//
-	// by Todd Gamblin, edited by Violet Giraffe.
-
-	void* array[50];
-	size_t size;
-
-	// Get void*s for all entries on the stack.
-	size = backtrace(array, 50);
-
-	// Print out all the frames to stderr.
-	if (sig == SIGSEGV)
-	{
-		// This branch should always be taken.
-		fprintf(stderr, "error: segmentation fault\n");
-	}
-	else
-	{
-		fprintf(stderr, "error: signal %d:\n", sig);
-	}
-	backtrace_symbols_fd(array, size, STDERR_FILENO);
-	exit(1);
-}
-
-#endif
+#include "util/Signals.h"
 
 using namespace std;
 using namespace stride;
@@ -76,10 +33,8 @@ using namespace TCLAP;
 /// Main program of the stride simulator.
 int main(int argc, char** argv)
 {
-	#if HANDLE_SIGNALS
 	// Set up a signal handler.
-	signal(SIGSEGV, handle_segfault);
-	#endif
+	stride::util::setup_segfault_handler();
 
 	int exit_status = EXIT_SUCCESS;
 	try {
